move sql serialization tests to kernel tests
